attempt to detect ime dpad failure and fall back to fakeDpad()
sadly this negates some of the performance benefits, but this seems to have fixed the problem where some menus lose the ability to be navigated with the ime.

the funny thing is it's not relying on fakeDpad() to fix that, but the side effect of changing the input focus. after that, the ime dpad magically works again.